projectTeams renamed to projectInvites
