+++ resources/projectInvites/projectInvitesController.js
