+++ server.js
@@ -19,7 +19,7 @@ const teamRouter = require('./resources/team/teamRouter');
 const teamMemberRouter = require('./resources/teamMember/teamMemberRouter');
 const inviteRouter = require('./resources/invite/inviteRouter');
 const exploreRouter = require('./resources/explore/exploreRouter');
-const projectTeamsRouter = require('./resources/projectTeams/projectTeamsRouter');
+const projectInvitesRouter = require('./resources/projectInvites/projectInvitesRouter');
 
 // ***************** MIDDLEWARE **************************
 
@@ -47,6 +47,6 @@ server.use('/api/v1/team', teamRouter);
 server.use('/api/v1/teamMember', teamMemberRouter);
 server.use('/api/v1/invite', inviteRouter);
 server.use('/api/v1/explore', exploreRouter);
-server.use('/api/v1/projectTeams', projectTeamsRouter);
+server.use('/api/v1/projectInvites', projectInvitesRouter);
 
 module.exports = server;
